+++ changelog.md
@@ -1,3 +1,15 @@
+### 1.9-RC7 - 2020-09-14
+#### Fixed
+- compress "hashed" field and hashed" input field from bit string to base64
+
+### 1.9-RC6 - 2020-08-21
+#### Fixed
+- ignore empty "resultFields" in ReadPatientToken
+
+### 1.9-RC5 - 2020-07-24
+#### Fixed
+- disable compression of hashedFiled from bit string to base64
+
 ### 1.9-RC1 - 2020-06-03
 #### Added
 - new CheckMatch endpoint: will retrieve the best match for the patient by using a Token of type "checkMatch"